zng_wgt_layer

Module popup

Popup widget.

Structs

DefaultStyle

W Popup default style.

POPUP

Popup service.

Popup

W An overlay container.

PopupCloseRequestedArgs

Arguments for POPUP_CLOSE_REQUESTED_EVENT.

Enums

ContextCapture

Defines if a Popup! captures the build/instantiation context.

PopupCloseMode

Optional parameter for POPUP_CLOSE_CMD.

PopupState

Identifies the lifetime state of a popup managed by POPUP.

Statics

ANCHOR_MODE_VAR

Popup anchor mode.

CLOSE_ON_FOCUS_LEAVE_VAR

If popup will close when it no longer contains the focused widget.

CONTEXT_CAPTURE_VAR

Popup context capture.

POPUP_CLOSE_CMD

Close the popup.

POPUP_CLOSE_REQUESTED_EVENT

Closing popup event.

STYLE_FN_VAR

Contextual style variable.

Functions

anchor_mode

P Defines the popup placement and size for popups open by the widget or descendants.

close_delay

P Delay awaited before actually closing when popup close is requested.

close_on_focus_leave

P Popup behavior when it loses focus.

context_capture

P Defines if the popup captures the local context to load in the popup context.

is_close_delaying

P If close was requested for this layered widget and it is just awaiting for the close_delay.

on_popup_close_requested

P Closing popup event.

on_pre_popup_close_requested

P Preview on_popup_close_requested event.

style_fn

P Extends or replaces the widget style.