Function zng_wgt_layer::popup::close_on_focus_leave

pub fn close_on_focus_leave(
    child: impl UiNode,
    close: impl IntoVar<bool>
) -> impl UiNode

P Popup behavior when it loses focus.

If true the popup will close itself, is true by default.

This property must be set on the widget that opens the popup or a parent, not the popup widget itself.

Sets the CLOSE_ON_FOCUS_LEAVE_VAR.