zng_wgt_layer

Struct AnchorOffset

pub struct AnchorOffset {
    pub place: Point,
    pub origin: Point,
}

Represents two points that position a layer widget with its anchor widget.

The place point is layout in the anchor widget bounds, the origin point is layout in the layer widget bounds, the layer widget is offset so that the origin point aligns with the place point.

Fields
place: Point

Point in the anchor widget.

origin: Point

Point in the layer widget.

Implementations

impl AnchorOffset

pub fn new(point: Point) -> Self

New place and origin points from same point.

pub fn in_top() -> Self

Layer widget is horizontally centered on the anchor widget and the top edge aligns.

pub fn in_bottom() -> Self

Layer widget is horizontally centered on the anchor widget and the bottom edge aligns.

pub fn in_left() -> Self

Layer widget is vertically centered on the anchor widget and the left edge aligns.

pub fn in_right() -> Self

Layer widget is vertically centered on the anchor widget and the right edge aligns.

pub fn in_top_left() -> Self

Layer widget top-left corner aligns with the anchor widget top-left corner.

pub fn in_top_right() -> Self

Layer widget top-right corner aligns with the anchor widget top-right corner.

pub fn in_bottom_left() -> Self

Layer widget bottom-left corner aligns with the anchor widget bottom-left corner.

pub fn in_bottom_right() -> Self

Layer widget bottom-right corner aligns with the anchor widget bottom-right corner.

pub fn center() -> Self

Layer widget is centered on the anchor widget.

pub fn out_top() -> Self

Layer widget is horizontally centered on the anchor widget and its bottom edge aligns with the anchors top edge.

pub fn out_bottom() -> Self

Layer widget is horizontally centered on the anchor widget and its top edge aligns with the anchors bottom edge.

pub fn out_left() -> Self

Layer widget is vertically centered on the anchor widget and its right edge aligns with the anchors left edge.

pub fn out_right() -> Self

Layer widget is vertically centered on the anchor widget and its left edge aligns with the anchors right edge.

pub fn out_top_left() -> Self

Layer widget bottom-right corner aligns with anchor widget top-left corner.

pub fn out_top_right() -> Self

Layer widget bottom-left corner aligns with anchor widget top-right corner.

pub fn out_bottom_left() -> Self

Layer widget top-right corner aligns with anchor widget bottom-left corner.

pub fn out_bottom_right() -> Self

Layer widget bottom-right corner aligns with anchor widget top-left corner.

pub fn out_top_in_left() -> Self

Layer widget bottom-left corner aligns with anchor widget top-left corner.

pub fn out_top_in_right() -> Self

Layer widget bottom-right corner aligns with anchor widget top-right corner.

pub fn out_bottom_in_left() -> Self

Layer widget top-left corner aligns with anchor widget bottom-left corner.

pub fn out_bottom_in_right() -> Self

Layer widget top-right corner aligns with anchor widget bottom-right corner.

pub fn out_left_in_top() -> Self

Layer widget top-right corner aligns with anchor widget top-left corner.

pub fn out_left_in_bottom() -> Self

Layer widget bottom-right corner aligns with anchor widget bottom-left corner.

pub fn out_right_in_top() -> Self

Layer widget top-left corner aligns with anchor widget top-right corner.

pub fn out_right_in_bottom() -> Self

Layer widget bottom-left corner aligns with anchor widget bottom-right corner.