Struct zng_wgt_layer::LayerIndex

pub struct LayerIndex(pub u32);

Represents a layer in a window.

See LAYERS for more information.

Tuple Fields
0: u32
Implementations

impl LayerIndex

pub const TOP_MOST: LayerIndex = _

The top-most layer.

Only widgets that are pretending to be a child window should use this layer, including menus, drop-downs, pop-ups and tooltips.

This is the u32::MAX value.

pub const ADORNER: LayerIndex = _

The layer for adorner display items.

Adorner widgets are related to another widget but not as a visual part of it, examples of adorners are resize handles in a widget visual editor, or an interactive help/guide feature.

This is the TOP_MOST - u16::MAX value.

pub const DEFAULT: LayerIndex = _

The default layer, just above the normal window content.

This is the 0 value.

pub fn saturating_add(self, other: impl Into<LayerIndex>) -> Self

Compute self + other saturating at the TOP_MOST bound instead of overflowing.

pub fn saturating_sub(self, other: impl Into<LayerIndex>) -> Self

Compute self - other saturating at the DEFAULT bound instead of overflowing.

pub fn name(self) -> Option<&'static str>

Gets the const name of this value.