Crate zng_wgt_layer

Window layers and popup.

Crate

This crate is part of the zng project.

Modules

popup

Popup widget.

Structs

AnchorMode

Defines what properties the layered widget takes from the anchor widget.

AnchorOffset

Represents two points that position a layer widget with its anchor widget.

LAYERS

Windows layers.

LayerIndex

Represents a layer in a window.

Enums

AnchorSize

Options for AnchorMode size constraints.

AnchorTransform

Options for AnchorMode::transform.

Functions

adorner

P Custom layered foreground.

adorner_fn

P Custom layered foreground generated using a WidgetFn<()>.

layers_node

Node that implements the layers, must be inserted in the NestGroup::EVENT group by the window implementer.